* async await update

+++ server/server.js
@@ -15,31 +15,30 @@ const port = process.env.PORT;
 
 app.use(bodyParser.json());
 
-app.post('/todos', authenticate, (req, res) => {
+app.post('/todos', authenticate, async (req, res) => {
     let todo = new Todo({
         text: req.body.text,
         _creator: req.user._id
     });
-
-    todo.save().then((doc)=>{
+    try {
+        let doc = await todo.save();
         res.send(doc);
-
-    }, (err)=>{
-        res.status(400).send(err);
-        })
-});
-
-app.get('/todos', authenticate, (req, res) => {
-    Todo.find({
-        _creator: req.user._id
-    }).then((todos) => {
-        res.send({todos});
-    },(e)=>{
+    } catch (e) {
         res.status(400).send(e);
-    })
+    }
 });
 
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos', authenticate, async (req, res) => {
+    try {
+        let todos = await Todo.find({_creator: req.user._id});
+        res.send({todos});
+    }
+    catch (e) {
+        res.status(400).send(e);
+    }
+});
+
+app.get('/todos/:id', authenticate, async (req, res) => {
 
    let id = req.params.id;
 
@@ -47,42 +46,36 @@ app.get('/todos/:id', authenticate, (req, res) => {
        return res.status(404).send();
    }
 
-    Todo.findOne({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
-       if(!todo){
+    try {
+        let todo = await Todo.findOne({_id: id, _creator: req.user._id});
+        if (!todo) {
            return res.status(404).send();
        }
-
        res.send({todo});
-   }).catch((e) => {
+    } catch (e) {
        res.status(400).send();
-   });
-
+    }
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    try {
+        let id = req.params.id;
 
-    if (!ObjectID.isValid(id)) {
-        return res.status(404).send();
-    }
+        if (!ObjectID.isValid(id)) {
+            return res.status(404).send();
+        }
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
+        let todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
         if (!todo) {
             return res.status(404).send();
         }
         res.send({todo});
-    }).catch((e) => {
+    } catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
     let id = req.params.id;
     let body = _.pick(req.body, ['text', 'completed']);
 
@@ -90,59 +83,60 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         return res.status(404).send();
     }
 
-    if(_.isBoolean(body.completed) && body.completed){
+    if (_.isBoolean(body.completed) && body.completed) {
         body.completedAt = new Date().getTime();
-    } else{
+    } else {
         body.completed = false;
         body.completedAt = null;
     }
 
-    Todo.findOneAndUpdate({_id: id, _creator: req.user._id}, {$set: body}, {new: true})
-        .then((todo) => {
-            if (!todo) {
+    try {
+        let todo = await Todo.findOneAndUpdate({_id: id, _creator: req.user._id}, {$set: body}, {new: true});
+        if (!todo) {
             return res.status(404).send();
         }
-            res.send({todo});
-        }).catch((e) => {
+        res.send({todo});
+    } catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
-app.post('/users', (req, res)=> {
-    let body = _.pick(req.body, ['email', 'password']);
-    let user = new User(body);
-
-    user.save().then(()=> user.generateAuthToken())
-        .then((token)=>{
-            res.header('x-auth', token).send(user);
-        })
-        .catch((e)=>{
+app.post('/users', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        let user = new User(body);
+        await user.save();
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    }
+    catch (e) {
         res.status(400).send(e);
-    });
-
+    }
 });
 
 app.get('/users/me', authenticate, (req, res)=>{
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password']);
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });
-    }).catch((e) => {
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        return res.header('x-auth', token).send(user);
+    }
+    catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token);
         res.status(200).send();
-    }, () => {
+    } catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
 app.listen(port, ()=> {
